Add support for generating numeric prefixes for page names
Confluence will auto-alphabetize all articles that are published
to it via the REST API. As far as I've been able to find, there
is no way to set the order through the API.

This is undesirable, as it causes some weird ordering errors when
using fine grained pagination modes - for ex: Definitions would
come before Overview, when Overview should obviously come first.

We can solve this by number prefixing any generated subpages,
which will cause confluence to retain the order. This has been
added as a configurable property 'generateNumericPrefixes'.

The default value is on. If you don't want numeric prefixes on
the subpages, you can turn it off - just be aware that
Confluence will then alpha-sort the pages, so the documentation
may be out of the desired order.