Prevent published pages from being orphaned
In this case that the user did not specify an ancestor id, the page
would be orphaned because confluence does not have a default.

When a page is orphaned, you cannot get to it unless you explictly
know the URL - it cannot be found in the space menu.

In this case, the code has been updated to probe the space's root
page and use that as the ancestor.