Allow toggling the Table of Contents on Single Page Mode
Previously Table of Contents was not available in the
single page mode, however, this provides support for it.

If you have more than and service or two but want single
page mode then this is very useful now that links are
translating correctly. However, if you have a very small
API it is just extra clutter, so I am keeping the option
to suppress it in this case.

Table of Contents is now on by default, but can be
suppress with includeTableOfContentsOnSinglePage = false