Individual Pagination: Do a "clean" deploy
At the fine grained level of pagination individual, pages
may need to be deleted - for example, if a service is removed.

This could also affected the number prefixes and thus the links
if you are using numeric prefixes

To work around these issues, we will do a clean build. Pages
at the individual level are deleted and recreated each deploy.

Confluence is queried for the subpages of the category pages,
and will clean them before recreating them.